pyproject.toml
[project]
name = "visa-dispute-advisor"
version = "0.1.0"
description = "MCP server for VISA Dispute Management Guidelines semantic search and merchant history lookup"
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=2.0",
    "chromadb>=0.6",
    "sqlalchemy>=2.0",
    "pdfplumber>=0.11",
    "polars>=1.0",
    "python-dotenv>=1.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "ruff>=0.5",
]

[project.scripts]
visa-advisor = "visa_dispute_advisor.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/visa_dispute_advisor"]

[tool.ruff]
target-version = "py311"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
